Add template and routes for ending a game (#40) (#45)
<!--- Please write your PR name in the present imperative tense.
Examples of that tense are: "Fix issue in the dispatcher where…",
"Improve our handling of…", etc." -->
<!-- For more information on Pull Requests, you can reference here:
https://success.vanillaforums.com/kb/articles/228-using-pull-requests-to-contribute
-->
## Describe Your Changes

Created template and get route for ending a game. Extracted some logic
into helper functions.

Outlined post request adding points to a players session. Currently I
don't see a place to store this information in the database. It's not
quiet what the Denar table calls for. I would add a table called
Sessions and SessionPlayers which covers the whole session and not just
a single game.

## Non-Obvious Technical Information


## Checklist Before Requesting a Review
- [x] The code runs successfully.

See issue #40 for a screen shot of the template in action.

---------

Co-authored-by: michplunkett <5885605+michplunkett@users.noreply.github.com>